# frontend/docs/tour_api_openapi_day1.yaml
openapi: 3.0.3
info:
  title: Tourista Tours API (Day 1 Contract)
  version: 1.0.0-day1
servers:
  - url: http://localhost:8080/api
paths:
  /tours:
    get:
      summary: Get active tours
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourSummaryListResponse"
  /tours/featured:
    get:
      summary: Get featured tours
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 6
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourSummaryListResponse"
  /tours/search:
    get:
      summary: Search tours
      parameters:
        - in: query
          name: city
          required: true
          schema:
            type: string
        - in: query
          name: departureDate
          schema:
            type: string
            format: date
        - in: query
          name: adults
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: children
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: durationMin
          schema:
            type: integer
        - in: query
          name: durationMax
          schema:
            type: integer
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [EASY, MEDIUM, HARD]
        - in: query
          name: minRating
          schema:
            type: number
        - in: query
          name: sort
          schema:
            type: string
            enum: [RECOMMENDED, PRICE_ASC, RATING_DESC, DEPARTURE_ASC]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourSummaryListResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tours/{id}:
    get:
      summary: Get tour detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/tours:
    post:
      summary: Create tour booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourBookingRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTourBookingResponse"
        "400":
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/my:
    get:
      summary: Get my bookings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBookingListResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiMeta:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    TourSummary:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        city: { type: string }
        categoryName: { type: string }
        durationDays: { type: integer }
        durationNights: { type: integer }
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        avgRating: { type: number }
        reviewCount: { type: integer }
        coverImage:
          type: string
          nullable: true
        pricePerAdult: { type: number }
        pricePerChild: { type: number }
        nearestDepartureDate:
          type: string
          format: date
          nullable: true
        availableSlots: { type: integer }
    TourSummaryListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiMeta"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/TourSummary"
    TourItineraryItem:
      type: object
      properties:
        dayNumber: { type: integer }
        title: { type: string }
        description: { type: string }
    TourDepartureItem:
      type: object
      properties:
        departureId: { type: integer }
        departureDate:
          type: string
          format: date
        availableSlots: { type: integer }
        priceOverride:
          type: number
          nullable: true
    TourDetailData:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        city: { type: string }
        categoryName: { type: string }
        description: { type: string }
        highlights:
          type: array
          items: { type: string }
        includes:
          type: array
          items: { type: string }
        excludes:
          type: array
          items: { type: string }
        durationDays: { type: integer }
        durationNights: { type: integer }
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        maxGroupSize: { type: integer }
        minGroupSize: { type: integer }
        avgRating: { type: number }
        reviewCount: { type: integer }
        coverImage:
          type: string
          nullable: true
        images:
          type: array
          items: { type: string }
        pricePerAdult: { type: number }
        pricePerChild: { type: number }
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/TourItineraryItem"
        departures:
          type: array
          items:
            $ref: "#/components/schemas/TourDepartureItem"
    TourDetailResponse:
      allOf:
        - $ref: "#/components/schemas/ApiMeta"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TourDetailData"
    CreateTourBookingRequest:
      type: object
      required: [tourId, departureId, adults]
      properties:
        tourId: { type: integer }
        departureId: { type: integer }
        adults:
          type: integer
          minimum: 1
        children:
          type: integer
          minimum: 0
          default: 0
        guestName:
          type: string
          maxLength: 100
        guestEmail:
          type: string
          maxLength: 150
        guestPhone:
          type: string
          maxLength: 20
        specialRequests:
          type: string
          maxLength: 1000
    CreateTourBookingData:
      type: object
      properties:
        bookingId: { type: integer }
        bookingCode: { type: string }
        status: { type: string }
        totalAmount: { type: number }
        currency: { type: string }
        tourId: { type: integer }
        tourTitle: { type: string }
        departureId: { type: integer }
        departureDate:
          type: string
          format: date
        adults: { type: integer }
        children: { type: integer }
        createdAt:
          type: string
          format: date-time
    CreateTourBookingResponse:
      allOf:
        - $ref: "#/components/schemas/ApiMeta"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CreateTourBookingData"
    MyBookingItem:
      type: object
      properties:
        bookingId: { type: integer }
        bookingCode: { type: string }
        bookingType:
          type: string
          enum: [HOTEL, TOUR]
        status: { type: string }
        totalAmount: { type: number }
        currency: { type: string }
        createdAt:
          type: string
          format: date-time
        tourId:
          type: integer
          nullable: true
        tourTitle:
          type: string
          nullable: true
        departureId:
          type: integer
          nullable: true
        departureDate:
          type: string
          format: date
          nullable: true
        adults:
          type: integer
          nullable: true
        children:
          type: integer
          nullable: true
    MyBookingListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiMeta"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/MyBookingItem"
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiMeta"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: string